// source/frontend/src/models/api/generated/ApiGen_Concepts_Contact.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Concepts_Organization } from './ApiGen_Concepts_Organization';
import { ApiGen_Concepts_Person } from './ApiGen_Concepts_Person';

// LINK: @backend/apimodels/Models/Concepts/Contact/ContactModel.cs
export interface ApiGen_Concepts_Contact {
  id: string | null;
  person: ApiGen_Concepts_Person | null;
  organization: ApiGen_Concepts_Organization | null;
}